pyproject.toml
[tool.poetry]
name = "feature_importance"
version = "0.1.0"
description = "machine learning feature importance techniques."
authors = ["christopher-cirelli <ccirelli2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-decouple = "^3.8"
pre-commit = "^3.5.0"
pylint = "^3.0.2"
pytest = "^7.4.3"
gitpython = "^3.1.40"
pytest-cov = "^4.1.0"
black = "^23.11.0"
pydantic = "1.10.13"
scikit-learn = "^1.4.1.post1"
lightgbm = "^4.3.0"
pandas = "^2.2.0"
pyarrow = "^15.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
